- correcion propiedades

+++ Restaurant.Model/Clases/IngredientePlato.cs
@@ -11,7 +11,7 @@
         }
 
         public int Id { get; set; }
-        public int CantidadInsumo { get; set; }
+        public double CantidadInsumo { get; set; }
         public int IdInsumo { get; set; }
         public int IdPlato { get; set; }
         public Insumo Insumo { get; set; }
